fix: Default to bf16 instead of fp16 for mixed precision
- Change default mixed precision from fp16 to bf16 when boolean True is given
- Aligns with modern GPU capabilities (RTX 5090 and similar)
- Better numerical stability than fp16

+++ src/neobert/glue/train.py
@@ -446,7 +446,7 @@ def trainer(cfg: Config):
     # Handle mixed precision setting (could be bool or string)
     mixed_precision = cfg.trainer.mixed_precision
     if isinstance(mixed_precision, bool):
-        mixed_precision = "no" if not mixed_precision else "fp16"
+        mixed_precision = "no" if not mixed_precision else "bf16"
     elif mixed_precision == "fp32":
         mixed_precision = "no"
